Hieu

+++ app/Http/Controllers/AdminToursController.php
@@ -2,6 +2,7 @@
 
 namespace App\Http\Controllers;
 
+use App\Models\orders_tours;
 use Illuminate\Http\Request;
 use App\Models\tours;
 use Illuminate\Support\Facades\Storage;
@@ -37,10 +38,10 @@ class AdminToursController extends Controller
         if ($request->hasFile('image')) {
             // Lưu ảnh vào thư mục public/images
             $imageName = time() . '-' . $request->file('image')->getClientOriginalName();
-            $request->file('image')->move(public_path('images'), $imageName);
+            $request->file('image')->move(public_path('images/upload/'), $imageName);
 
             // Lưu đường dẫn tương đối của ảnh
-            $imagePath = 'images/upload' . $imageName;
+            $imagePath = 'images/upload/' . $imageName;
         }
 
         // Tạo tour mới với đường dẫn ảnh
@@ -63,9 +64,21 @@ class AdminToursController extends Controller
     }
 
     // Hiển thị danh sách các tours
-    public function index()
+    public function index(Request $request)
     {
-        $tours = tours::all();
+        // Tạo query để có thể áp dụng các điều kiện tìm kiếm và sắp xếp
+    $query = tours::query();
+
+    // Kiểm tra và áp dụng tìm kiếm theo tên
+    if ($request->has('name') && $request->name != '') {
+        $query->where('name', 'LIKE', '%' . $request->name . '%');
+    }
+
+    // Sắp xếp theo cột 'created_at' từ mới đến cũ
+    $query->orderBy('created_at', 'desc');
+
+        // Phân trang 15 user mỗi trang
+        $tours = $query->paginate(15);  
         return view('admin.tours.index', compact('tours'));
     }
 
@@ -109,7 +122,7 @@ public function update(Request $request, $id)
 
         // Lưu ảnh mới vào thư mục public/images
         $imageName = time() . '-' . $request->file('image')->getClientOriginalName();
-        $request->file('image')->move(public_path('images'), $imageName);
+        $request->file('image')->move(public_path('images/upload/'), $imageName);
 
         // Lưu đường dẫn tương đối của ảnh
         $tour->image = 'images/upload/' . $imageName;
@@ -140,6 +153,11 @@ public function destroy($id)
     // Tìm tour theo id
     $tour = tours::findOrFail($id);
 
+    if (orders_tours::where('tour_id', $tour->id)->exists()) {
+        // Chuyển hướng với thông báo lỗi
+        return redirect()->route('admin.tours.index')->with('error', 'This tour cannot be deleted because it has been booked.');
+    }
+    
     // Xóa ảnh trong thư mục public/images (nếu có)
     if ($tour->image && file_exists(public_path($tour->image))) {
         unlink(public_path($tour->image));
